Big Changes, Working CLI (somewhat)
got it kinda working.
when run it returns the player name + tag, rank and RR. the level and peak rank are still to be implemented as I am trying to find the endpoint and whatnot.

Added some config stuff. currently not used but will be when I actually get the data properly returned.

the following files are not being used but will be fully implemented when I get this darn thing working:
src/utils/ArgumentHandler.js
src/utils/ConfigManager.js
config/default.json